requirements.txt
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.4
Pillow==10.1.0
scikit-image==0.22.0
scipy==1.11.4
sympy==1.12
torch==2.1.1
tqdm==4.66.1
